combine category and entry repository
